+++ philosophers/utils1.c
@@ -6,16 +6,16 @@
 /*   By: yachen <yachen@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/09/13 14:13:35 by yachen            #+#    #+#             */
-/*   Updated: 2023/10/10 20:38:34 by yachen           ###   ########.fr       */
+/*   Updated: 2023/10/11 12:36:58 by yachen           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include  "./includes/philosophers.h"
 
-int	philo_ft_atoi(char *str)
+long	philo_ft_atoi(char *str)
 {
-	int		i;
-	int		nb;
+	int			i;
+	long		nb;
 
 	i = 0;
 	nb = 0;
@@ -38,19 +38,18 @@ void	clean_forks(pthread_mutex_t **fork, int nb_fork)
 		pthread_mutex_destroy(&(*fork)[i++]);
 }
 
+/* Clean ressources before exit program */
 void	clean_all(t_pgm *pgm)
 {
 	pthread_mutex_destroy(&pgm->write_lock);
-//	clean_forks(&pgm->meal_lock, pgm->philos[0].nb_of_philos);
 	pthread_mutex_destroy(&pgm->dead_lock);
 	pthread_mutex_destroy(&pgm->meal_lock);
 	clean_forks(&pgm->forks, pgm->philos[0].nb_of_philos);
-//	free(pgm->meal_lock);
 	free(pgm->forks);
 	free(pgm->philos);
 }
 
-// Gets the current time in milliseconds
+/* Gets the current time in milliseconds */
 size_t	get_current_time(void)
 {
 	struct timeval	time;
